requirements.txt
aniso8601==3.0.2
Django==2.2.1
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.17.6
django-haystack==2.8.1
django-nine==0.1.13
djangorestframework==3.9.3
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
ipaddress==1.0.22
promise==2.2.1
python-dateutil==2.8.0
pytz==2019.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.2
